Count extremely easy extended backends (3ms) as regular backends
If something is off by a few ticks, numerically a backend may be a few ms after the end of sp but it's basically a regular backend. Based off the case of Missed Injections whose path has a fill that's early by a couple of ticks, which is imperceptible but makes the backend technically an extended one.

Also made some edits to Run Folder since Record formats changed, still a work in progress