Activation misalignment accuracy for .mid
Activation misalignment is when the end of an activation phrase doesn't coincide with a chord. We have to make a nearby chord the activation chord.

In some cases it's the previous chord, in other cases it's the upcoming chord.

Fixes #4.

Also removed the "alldata" tests since they're too broad and will break easily