SqIn/SqOut cleanup and difficulty displays
Reworked SqIn/SqOut data. Paths and activations now have a concept of "difficulty" measured in milliseconds, based on the hardest SqIn/SqOut/E that they contain.

In the app, those timings for SqIn and SqOut can be viewed on activations and paths, and difficult ones are given a warning color.

Resolves #22