Mult Squeeze update
Mult squeezes are no longer a side calculation during the scoring of each chord. Now there's just a more specific check for whether a mult squeeze is possible on each chord. Also cleaned up their representation based on the knowledge that

With all that, a more helpful "how to hit" message is shown in the app for each mult squeeze.

Also added some support stuff for Chord and some testing.

Closes #21