Don't compare paths that recently SqIn/SqOuted
The way SqIn/SqOuts are currently handled, paths can be in a state where they've handled an SP phrase early as part of carrying out the SqIn/SqOut, and are ready to ignore that SP phrase when it normally occurs right after. In this split second the SqIn/SqOut paths have an extra SP compared to paths that are about to gain it normally, allowing them to seem like better paths even when they're not.

Fixed this by, in the path elimination step, ignoring paths that are in this temporary buffered_sqinout_sp state.

Fixes #40