E calculation accuracy update
E is based on the length of the fill in realtime rather than the length of the fill in ticks. Also there's a 1/64th note of padding that makes E thresholds slightly before the note you'd expect.

Flipped the sign of e offsets in the app after having some more practical experience comparing it to the in-game note accuracy

Added basic tests for e offsets changing with bpm, and a couple of songs that had been tricky.

Manually reviewed the sb25 regressions, some activations became trivial Es without changing the path, which is expected.

Resolves #44