Timecode records optimization
Saved timecodes as just the tick value instead of the recursive object with tick, mbt, and ms.

In order to rebuild timecodes from just their tick value, the song's tick resolution, bpms, and tpms (time signatures) are now saved. The cost of this is massively outweighed by the reduced footprint of timecodes.