Variants update
A record's paths now consist of the list of base paths and nested variants.
Variants are prepped after analysis and after loading, instead of once after analysis (after which the variant system is neutralized).
Privated a record's paths to encourage using all_paths(), abstracting away getting nested variants
A path's activations now consist of the main activations list and a variant tail which is copied from the parent path (or empty if the path is not a variant). The variant tail is not saved to json since it's redundant.
Privated a path's activations to encourage using all_activations(), abstracting away the combination of _activations and _variant_tail
Dummied out record's __eq__ because it lacks a purpose right now
Added all_paths, best_path, all_activations, get_activation, has_activations, and __len__ to improve use cases that previously accessed a record's paths or a path's activations directly
Resolves #50